refactory: strategy 관련 코드 패키지 변경
