+++ src/main/java/blackjack/domain/card/Deck.java
@@ -1,5 +1,9 @@
 package blackjack.domain.card;
 
+import blackjack.domain.card.strategy.DrawCardsStrategy;
+import blackjack.domain.card.strategy.RandomShuffleStrategy;
+import blackjack.domain.card.strategy.ShuffleStrategy;
+
 import java.util.List;
 import java.util.Stack;
