+++ src/test/java/blackjack/domain/card/DeckTest.java
@@ -1,5 +1,6 @@
 package blackjack.domain.card;
 
+import blackjack.domain.card.strategy.InitialCardsDrawStrategy;
 import org.junit.jupiter.api.*;
 
 import java.util.List;
